requirements.txt
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
torchmetrics==1.8.2
transformers==4.56.2
ultralytics==8.3.202
ultralytics-thop==2.0.17
uri-template==1.3.0
urllib3
accelerate==1.10.1
peft==0.17.1
timm==1.0.19
scikit-learn==1.7.1
numpy==2.2.6
pandas==2.3.2
scipy==1.15.3
open_clip_torch==3.1.0
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
Pillow==10.3.0
scikit-image==0.25.2
matplotlib==3.9.0
seaborn==0.13.2
notebook_shim==0.2.4
nbconvert==7.16.4
nbformat==5.10.4
huggingface-hub==0.35.0
safetensors==0.6.2
tqdm==4.66.4
PyYAML==6.0.1
requests==2.32.3

diffusers==0.35.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
wandb==0.22.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
widgetsnbextension==4.0.11
zipp==3.23.0
clip @ git+https://github.com/openai/CLIP.git